+++ CHANGELOG.md
@@ -158,6 +158,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `jj workspace forget` now warns about unknown workspaces instead of failing.
 
+* Added a new commit signing backend for Sequoia PGP.
+
 ### Fixed bugs
 
 * Broken symlink on Windows. [#6934](https://github.com/jj-vcs/jj/issues/6934).